haba_data_base rename apps folders uninstall skeditor add comment to Readme - django-summernote
