+++ README.md
@@ -2,7 +2,7 @@
 Командная разработка по методологии Agile: SCRUM
 
 Django Modified Preorder Tree Traversal: https://pypi.org/project/django-mptt/ \
-Django admin CKEditor integration: https://pypi.org/project/django-ckeditor/ \
 Django Debug Toolbar: https://pypi.org/project/django-debug-toolbar/ \
 Django environ: https://pypi.org/project/django-environ/ \
 Pillow: https://pypi.org/project/Pillow/
+https://pypi.org/project/django-summernote/ - Редактор текстового поля
