+++ haba_haba/haba_haba/settings.py
@@ -46,12 +46,10 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'ckeditor',
-    'ckeditor_uploader',
     'django_summernote',
     'debug_toolbar',
-    'users.apps.UsersConfig',
-    'haba_app.apps.HabaAppConfig',
+    'userapp.apps.UsersConfig',
+    'mainapp.apps.HabaAppConfig',
 ]
 
 MIDDLEWARE = [
@@ -149,7 +147,7 @@ STATICFILES_DIRS = [
 MEDIA_URL = 'media/'
 MEDIA_ROOT = BASE_DIR / 'media/'
 
-AUTH_USER_MODEL = 'users.HabaUser'
+AUTH_USER_MODEL = 'userapp.HabaUser'
 
 # Summernote
 SUMMERNOTE_THEME = 'bs4'
