+++ haba_haba/mainapp/apps.py
@@ -3,5 +3,5 @@ from django.apps import AppConfig
 
 class HabaAppConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'haba_app'
+    name = 'mainapp'
     verbose_name = 'HABA-HABA'
