+++ haba_haba/userapp/apps.py
@@ -3,5 +3,5 @@ from django.apps import AppConfig
 
 class UsersConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'users'
+    name = 'userapp'
     verbose_name = "Пользователи"
